Disable checkbox while scrip loading
To prevent you from clicking on the checkbox
while the PreviewScrips widget is loading,
I've overridden the click handler to return false
preventing the click event from working.

It should be noted that "disabled" and "readonly"
attributres do not have the desired behavior on
checkboxes as they impact the value and not the
check state.

Fixes: I#32290